refactor: rename mod and extract bit_torrent mod
`protocol` nod contains logic not only for BitTorrent protocol.
