--- src/protocol/mod.rs
@@ -1,5 +1,0 @@
-pub mod clock;
-pub mod common;
-pub mod crypto;
-pub mod info_hash;
-pub mod utils;
