+++ src/servers/http/v1/handlers/scrape.rs
@@ -99,10 +99,10 @@ mod tests {
 
     use torrust_tracker_test_helpers::configuration;
 
-    use crate::protocol::info_hash::InfoHash;
     use crate::servers::http::v1::requests::scrape::Scrape;
     use crate::servers::http::v1::responses;
     use crate::servers::http::v1::services::peer_ip_resolver::ClientIpSources;
+    use crate::shared::bit_torrent::info_hash::InfoHash;
     use crate::tracker::services::common::tracker_factory;
     use crate::tracker::Tracker;
