+++ src/shared/bit_torrent/mod.rs
@@ -1,0 +1,2 @@
+pub mod common;
+pub mod info_hash;
