+++ src/tracker/databases/mysql.rs
@@ -10,8 +10,8 @@ use r2d2_mysql::MySqlConnectionManager;
 use torrust_tracker_primitives::DatabaseDriver;
 
 use super::{Database, Error};
-use crate::protocol::common::AUTH_KEY_LENGTH;
-use crate::protocol::info_hash::InfoHash;
+use crate::shared::bit_torrent::common::AUTH_KEY_LENGTH;
+use crate::shared::bit_torrent::info_hash::InfoHash;
 use crate::tracker::auth::{self, Key};
 
 const DRIVER: DatabaseDriver = DatabaseDriver::MySQL;
