+++ tests/common/fixtures.rs
@@ -1,7 +1,7 @@
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 
 use aquatic_udp_protocol::{AnnounceEvent, NumberOfBytes};
-use torrust_tracker::protocol::clock::DurationSinceUnixEpoch;
+use torrust_tracker::shared::clock::DurationSinceUnixEpoch;
 use torrust_tracker::tracker::peer::{self, Id, Peer};
 
 pub struct PeerBuilder {
